Use 32-bit sample clock and const resonator probe

+++ src/hal/hal_audio.cpp
@@ -22,7 +22,7 @@ void *g_user_data = nullptr;
 bool g_running = false;
 float g_sample_rate = 44100.0f;
 std::size_t g_frames_per_block = 64;
-std::atomic<uint64_t> g_sample_clock{0};
+std::atomic<uint32_t> g_sample_clock{0};
 
 #ifdef SEEDBOX_HW
 class CallbackStream : public AudioStream {
@@ -69,7 +69,7 @@ class CallbackStream : public AudioStream {
     transmit(right_block, 1);
     release(left_block);
     release(right_block);
-    g_sample_clock.fetch_add(AUDIO_BLOCK_SAMPLES, std::memory_order_relaxed);
+    g_sample_clock.fetch_add(static_cast<uint32_t>(AUDIO_BLOCK_SAMPLES), std::memory_order_relaxed);
   }
 };
 
@@ -143,7 +143,7 @@ void mockPump(std::size_t frames) {
   // so unit tests exercise the same logic.
   StereoBufferView view{scratch(left, frames).data(), scratch(right, frames).data(), frames};
   g_callback(view, g_user_data);
-  g_sample_clock.fetch_add(frames, std::memory_order_relaxed);
+  g_sample_clock.fetch_add(static_cast<uint32_t>(frames), std::memory_order_relaxed);
 }
 #endif  // SEEDBOX_HW
